+++ build-tools/find_latest_hard_migration_base_version.sh
@@ -8,5 +8,8 @@
 
 set -euo pipefail
 
-latest_release=$(cat "$SPLICE_ROOT/LATEST_RELEASE")
-echo "release-line-${latest_release}"
+# latest_release=$(cat "$SPLICE_ROOT/LATEST_RELEASE")
+# echo "release-line-${latest_release}"
+
+# TODO(DACH-NY/canton-network-internal#2180): revert this to release-line-${latest_release} once latest_release is 0.4.21
+echo "0.4.20-with-auth0-changes"